requirements.txt
flask
firebase_admin
gunicorn
python-dotenv
google-generativeai
APScheduler
cloudinary
